# api/docs/swagger.yaml
definitions:
  models.ResponseError:
    properties:
      error: {}
    type: object
  models.ResponseSuccess:
    properties:
      data: {}
      metadata: {}
    type: object
  user_service.BranchPrimaryKey:
    properties:
      id:
        type: string
    type: object
  user_service.CreateBranch:
    properties:
      active:
        type: boolean
      address:
        description: 'Note: Corrected spelling to "address"'
        type: string
      close_time:
        type: string
      location:
        allOf:
        - $ref: '#/definitions/user_service.Location'
        description: 'Note: Changed to lowercase "location"'
      name:
        type: string
      open_time:
        type: string
      phone:
        type: string
    type: object
  user_service.CreateCustomer:
    properties:
      date_of_birth:
        type: string
      gender:
        type: string
      gmail:
        type: string
      language:
        type: string
      phone:
        type: string
    type: object
...
  user_service.CustomerPrimaryKey:
    properties:
      id:
        type: string
    type: object
  user_service.GetListBranchRequest:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      search:
        type: string
    type: object
  user_service.GetListCustomerRequest:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      search:
        type: string
    type: object
...
  user_service.Location:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
...
  user_service.UpdateBranchRequest:
    properties:
      active:
        type: boolean
      address:
        description: 'Note: Corrected spelling to "address"'
        type: string
      close_time:
        type: string
      id:
        type: string
      location:
        allOf:
        - $ref: '#/definitions/user_service.Location'
        description: 'Note: Changed to lowercase "location"'
      name:
        type: string
      open_time:
        type: string
      phone:
        type: string
    type: object
  user_service.UpdateCustomerRequest:
    properties:
      date_of_birth:
        type: string
      gender:
        type: string
      gmail:
        type: string
      id:
        type: string
      language:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/branch/create:
    post:
      consumes:
      - application/json
      description: This api creates a Branch and returns its id
      parameters:
      - description: Branch
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Creates a Branch
      tags:
      - Branch
  /v1/branch/delete:
    delete:
      description: This api delete a Branch
      parameters:
      - description: Branch
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/user_service.BranchPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: delete a Branch
      tags:
      - Branch
  /v1/branch/getbyid:
    post:
      description: This api get by id a Branch
      parameters:
      - description: Branch
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/user_service.BranchPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get by id a Branch
      tags:
      - Branch
  /v1/branch/getlist:
    post:
      description: This api get list a Branch
      parameters:
      - description: Branch
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/user_service.GetListBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get list a Branch
      tags:
      - Branch
  /v1/branch/updatebranch:
    put:
      consumes:
      - application/json
      description: This API updates a Branch
      parameters:
      - description: Branch object to update
        in: body
        name: Branch
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update a Branch
      tags:
      - Branch
  /v1/customer/create:
    post:
      consumes:
      - application/json
      description: This api creates a customer and returns its id
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateCustomer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Creates a customer
      tags:
      - Customer
  /v1/customer/delete:
    delete:
      description: This api delete a customer
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/user_service.CustomerPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: delete a customer
      tags:
      - Customer
  /v1/customer/getbyid:
    post:
      description: This api get by id a customer
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/user_service.CustomerPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get by id a customer
      tags:
      - Customer
  /v1/customer/getlist:
    post:
      description: This api get list a customer
      parameters:
      - description: customer
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/user_service.GetListCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get list a customer
      tags:
      - Customer
  /v1/customer/update:
    put:
      consumes:
      - application/json
      description: This API updates a customer
      parameters:
      - description: Customer object to update
        in: body
        name: customer
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update a customer
      tags:
      - Customer
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"